refactor code style.
